# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Chainmail – Unified open-source maintainer workbench API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analyses/{id}/sbom:
    get:
      operationId: getSbomSummary
      tags: [sbom]
      summary: SBOM health summary for an analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: SBOM summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SbomSummary"
...
  /analyses/{id}/risk:
    get:
      operationId: getRiskScore
      tags: [risk]
      summary: Supply chain risk score breakdown
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Risk score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskScore"
...
components:
  schemas:
...
    SbomSummary:
      type: object
      properties:
        analysisId:
          type: string
        totalComponents:
          type: integer
        directDependencies:
          type: integer
        transitiveDependencies:
          type: integer
        staleComponents:
          type: integer
        uniqueLicenses:
          type: integer
        format:
          type: string
        formatVersion:
          type: string
        components:
          type: array
          items:
            $ref: "#/components/schemas/Component"
      required:
        - analysisId
        - totalComponents
        - directDependencies
        - transitiveDependencies
        - staleComponents
        - uniqueLicenses
        - format
        - formatVersion
        - components
    Component:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        version:
          type: string
        license:
          type: string
        isDirect:
          type: boolean
        isStale:
          type: boolean
        publishedAt:
          type: string
          format: date-time
        latestVersion:
          type: string
        vulnerabilityCount:
          type: integer
      required:
        - id
        - name
        - version
        - license
        - isDirect
        - isStale
        - vulnerabilityCount
...
    RiskScore:
      type: object
      properties:
        analysisId:
          type: string
        overallScore:
          type: number
        grade:
          type: string
          enum: [A, B, C, D, F]
        dimensions:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              score:
                type: number
              weight:
                type: number
              status:
                type: string
                enum: [good, warning, critical]
              detail:
                type: string
            required: [name, score, weight, status, detail]
        slsaLevel:
          type: integer
        recommendations:
          type: array
          items:
            type: string
      required:
        - analysisId
        - overallScore
        - grade
        - dimensions
        - slsaLevel
        - recommendations
...